backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
ffmpeg-python
requests
numpy
scipy
opencv-python-headless
pydub
soundfile
librosa
replicate
python-dotenv